+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add ols(setting) and wls(setting, weights = w) type method call where setting is a regression setting 
 - Implement cooksoutliers() method for determining potential regression outliers using a cutoff value.
 - Update documentation
+- Theil-Sen estimator for multiple regression
 
 
 # v0.9.1